LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-constructor-expected.txt

Testing the IOSInlineMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.classList.contains('media-controls') is true
PASS mediaControls.element.classList.contains('inline') is true
PASS mediaControls.element.classList.contains('ios') is true
PASS mediaControls.layoutTraits is LayoutTraits.iOS
PASS accessibilityController.accessibleElementById('groupping-label').description is "AXLabel: Video Controls"
PASS accessibilityController.accessibleElementById('groupping-label').description is "AXLabel: Audio Controls"

PASS successfullyParsed is true

TEST COMPLETE
